+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/SearchRepositoryDiffCallback.kt
@@ -1,0 +1,19 @@
+package jp.co.yumemi.android.code_check
+
+import androidx.recyclerview.widget.DiffUtil
+
+object SearchRepositoryDiffCallback : DiffUtil.ItemCallback<RepositoryInfo>() {
+    override fun areItemsTheSame(
+        oldItem: RepositoryInfo,
+        newItem: RepositoryInfo
+    ): Boolean {
+        return oldItem.name == newItem.name
+    }
+
+    override fun areContentsTheSame(
+        oldItem: RepositoryInfo,
+        newItem: RepositoryInfo
+    ): Boolean {
+        return oldItem == newItem
+    }
+}
