+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/SearchRepositoryFragment.kt
@@ -4,11 +4,8 @@
 package jp.co.yumemi.android.code_check
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.view.inputmethod.EditorInfo
-import android.widget.TextView
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.viewModelScope
@@ -17,7 +14,6 @@ import androidx.recyclerview.widget.*
 import jp.co.yumemi.android.code_check.databinding.SearchRepositoryFragmentBinding
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.launch
-import java.lang.Exception
 
 // TODO: MVVMを導入してViewModelに処理を移す
 /**
@@ -39,8 +35,8 @@ class SearchRepositoryFragment : Fragment(R.layout.search_repository_fragment) {
         val dividerItemDecoration =
             DividerItemDecoration(requireContext(), layoutManager.orientation)
         val adapter =
-            CustomAdapter(
-                object : CustomAdapter.OnItemClickListener {
+            RepositoryListAdapter(
+                object : RepositoryListAdapter.OnItemClickListener {
                     override fun itemClick(item: RepositoryInfo) {
                         gotoRepositoryFragment(item)
                     }
@@ -85,50 +81,5 @@ class SearchRepositoryFragment : Fragment(R.layout.search_repository_fragment) {
     override fun onDestroyView() {
         super.onDestroyView()
         _binding = null
-    }
-}
-
-// FIX: privateにする
-val diff_util =
-    object : DiffUtil.ItemCallback<RepositoryInfo>() {
-        override fun areItemsTheSame(
-            oldItem: RepositoryInfo,
-            newItem: RepositoryInfo
-        ): Boolean {
-            return oldItem.name == newItem.name
-        }
-
-        override fun areContentsTheSame(
-            oldItem: RepositoryInfo,
-            newItem: RepositoryInfo
-        ): Boolean {
-            return oldItem == newItem
-        }
-    }
-
-/**
- * 検索にヒットしたGitHubリポジトリをリスト表示するための
- * ListAdapterを継承したカスタムアダプター
- * */
-class CustomAdapter(
-    private val itemClickListener: OnItemClickListener,
-) : ListAdapter<RepositoryInfo, CustomAdapter.ViewHolder>(diff_util) {
-
-    class ViewHolder(view: View) : RecyclerView.ViewHolder(view)
-
-    interface OnItemClickListener {
-        fun itemClick(item: RepositoryInfo)
-    }
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-        val view = LayoutInflater.from(parent.context).inflate(R.layout.layout_item, parent, false)
-        return ViewHolder(view)
-    }
-
-    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        val item = getItem(position)
-        (holder.itemView.findViewById<View>(R.id.repositoryNameView) as TextView).text = item.name
-
-        holder.itemView.setOnClickListener { itemClickListener.itemClick(item) }
     }
 }
